pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[project]
name = "sam-adapter-namibia"
version = "0.1.0"
description = ""
authors = [
    {name = "yuming-p",email = "panpanpandora@hotmail.com"}
]
readme = "README.md"
requires-python = "==3.10.1"



dependencies = [
    "imageio",
    "ipython>=8.11.0",
    "matplotlib",
    "opencv-python>=4.7.0",
    "PyYAML==6.0",
    "scikit-learn",
    "scipy>=1.10.1",
    "tqdm>=4.51.0",
    "numpy>=1.20.3",
    "typing",
    "terminaltables",
    "Pillow",
    "tensorboardx (>=2.6.2.2,<3.0.0.0)",
    "onnxruntime (>=1.21.0,<2.0.0)",
    "setuptools (>=76.0.0,<77.0.0)",
    "timm (>=1.0.15,<2.0.0)",
    "easydict (>=1.13,<2.0)",
    "attr (>=0.3.2,<0.4.0)",
    "thop (>=0.1.1.post2209072238,<0.2.0)",
    "torchsummary (>=1.5.1,<2.0.0)",
    "rasterio (>=1.4.3,<2.0.0)",
    "imagecodecs (>=2024.12.30,<2025.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "mmcv (==1.7.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
    "shapely (>=2.0.7,<3.0.0)",
    "pyproj (>=3.7.1,<4.0.0)",
    "geopandas (>=1.0.1,<2.0.0)",
    "tensorboard (>=2.19.0,<3.0.0)",
    "fiona (>=1.10.1,<2.0.0)",
    "gdal (>=3.10.3,<4.0.0)",
    "black (>=25.1.0,<26.0.0)",
]


[tool.poetry]
package-mode = false
